backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.0
drf-spectacular>=0.26.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=5.0.0
django-storages>=1.14.0
Pillow>=10.0.0
whitenoise>=6.6.0
django-htmx>=1.18.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
django-extensions>=3.2.0
requests>=2.31.0
pymongo>=4.6.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
ruff>=0.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

